[meson compile clippy]

+++ src/renderer/src/luagfx.rs
@@ -137,6 +137,7 @@ impl UserData for LuaGfx {
          *    @luatparam[opt] number angle Angle to rotate in radians.
          * @luafunc renderTexRaw
          */
+        #[allow(clippy::type_complexity)]
         methods.add_function(
             "renderTex",
             |_,
@@ -165,28 +166,28 @@ impl UserData for LuaGfx {
                 #[rustfmt::skip]
                 let transform: Matrix3<f32> = {
                     if let Some(angle) = angle {
-                        let hw = 0.5 * w as f32;
-                        let hh = 0.5 * h as f32;
-                        let c = angle.cos() as f32;
-                        let s = angle.sin() as f32;
+                        let hw = 0.5 * w;
+                        let hh = 0.5 * h;
+                        let c = angle.cos();
+                        let s = angle.sin();
                         Matrix3::new(
-                            1.0, 0.0, x as f32 + hw,
-                            0.0, 1.0, y as f32 + hh,
+                            1.0, 0.0, x + hw,
+                            0.0, 1.0, y + hh,
                             0.0, 0.0, 1.0,
                         ) * Matrix3::new(
-                        c,  -s,  0.0,
-                        s,   c,  0.0,
+                        c,   -s,  0.0,
+                        s,    c,  0.0,
                         0.0, 0.0, 1.0,
                         ) * Matrix3::new(
-                        w as f32, 0.0,      -hw,
-                        0.0,      h as f32, -hh,
-                        0.0,      0.0,      1.0,
+                        w,   0.0, -hw,
+                        0.0,  h,  -hh,
+                        0.0, 0.0, 1.0,
                         )
                     } else {
                         Matrix3::new(
-                            w as f32, 0.0,      x as f32,
-                            0.0,      h as f32, y as f32,
-                            0.0,      0.0,      1.0,
+                            w,   0.0,  x,
+                            0.0,  h,   y,
+                            0.0, 0.0, 1.0,
                         )
                     }
                 };
